backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.0
pytest==7.4.0
httpx==0.24.1
bcrypt==4.0.1
python-dotenv==1.0.0
asyncpg==0.28.0  # PostgreSQL driver (if used)
aiosqlite==0.19.0  # SQLite driver for development
websockets==11.0.3  # For real-time notifications 